% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_arrivals}
\alias{calc_arrivals}
\title{Calculate the number of arrivals}
\usage{
calc_arrivals(arrivals, groups = NULL)
}
\arguments{
\item{arrivals}{Dataframe with times for each patient with each resource.}

\item{groups}{Optional list of columns to group by for the calculation.}
}
\value{
Tibble with column containing total number of arrivals.
}
\description{
Calculate the number of arrivals
}
