% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_mean_patients_in_service}
\alias{calc_mean_patients_in_service}
\title{Calculate the time-weighted mean number of patients in the service.}
\usage{
calc_mean_patients_in_service(patient_count, groups = NULL)
}
\arguments{
\item{patient_count}{Dataframe with patient counts over time.}

\item{groups}{Optional list of columns to group by for the calculation.}
}
\value{
Tibble with column containing mean number of patients in the service.
}
\description{
Calculate the time-weighted mean number of patients in the service.
}
