% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_results.R
\name{calc_unseen_mean}
\alias{calc_unseen_mean}
\title{Calculate the mean wait time of patients who are still waiting for a
resource at the end of the simulation}
\usage{
calc_unseen_mean(arrivals, groups = NULL)
}
\arguments{
\item{arrivals}{Dataframe with times for each patient with each resource.}

\item{groups}{Optional list of columns to group by for the calculation.}
}
\value{
Tibble with columns containing result for each resource.
}
\description{
Calculate the mean wait time of patients who are still waiting for a
resource at the end of the simulation
}
