% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_all_param_names}
\alias{check_all_param_names}
\title{Validate all parameter groups in a parameter list.}
\usage{
check_all_param_names(param)
}
\arguments{
\item{param}{List of parameter groups.}
}
\value{
Throws an error if any group has missing or extra parameters.
}
\description{
This function checks that the top-level and all nested parameter lists have
the correct names.
}
