% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_all_positive}
\alias{check_all_positive}
\title{Check if all values are positive}
\usage{
check_all_positive(x, name)
}
\arguments{
\item{x}{The vector or list of values to check.}

\item{name}{The name of the parameter (for error messages).}
}
\value{
None. Throws an error if the check fails.
}
\description{
Throws an error if any value in the input is not positive (> 0).
}
