% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_allowed_params}
\alias{check_allowed_params}
\title{Check that a parameter list contains only allowed names}
\usage{
check_allowed_params(object_name, actual_names, allowed_names)
}
\arguments{
\item{object_name}{String name of object (for error messages).}

\item{actual_names}{Character vector of parameter names.}

\item{allowed_names}{Character vector of allowed parameter names.}
}
\value{
None. Throws an error if unrecognised parameters are present.
}
\description{
Check that a parameter list contains only allowed names
}
