% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_log_file_path}
\alias{check_log_file_path}
\title{Check logging file path}
\usage{
check_log_file_path(param)
}
\arguments{
\item{param}{List containing parameters for the simulation.}
}
\value{
None. Throws an error if valid file path not provided, when
log_to_file = TRUE.
}
\description{
Check logging file path
}
