% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_nonneg_integer}
\alias{check_nonneg_integer}
\title{Check if a value is a non-negative integer}
\usage{
check_nonneg_integer(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{The name of the parameter (for error messages).}
}
\value{
None. Throws an error if the check fails.
}
\description{
Throws an error if the value is not a non-negative integer (>= 0).
}
