% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_param_names}
\alias{check_param_names}
\title{Validate parameter names.}
\usage{
check_param_names(valid_names, input_names)
}
\arguments{
\item{valid_names}{Character vector of valid argument names.}

\item{input_names}{Character vector of input parameter names.}
}
\value{
Throws an error if there are missing or extra parameters.
}
\description{
Ensure that all required parameters are present, and no extra parameters are
provided.
}
