% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_positive_integer}
\alias{check_positive_integer}
\title{Check if a value is a positive integer}
\usage{
check_positive_integer(x, name)
}
\arguments{
\item{x}{The value to check.}

\item{name}{The name of the parameter (for error messages).}
}
\value{
None. Throws an error if the check fails.
}
\description{
Throws an error if the value is not a positive integer (> 0).
}
