% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_prob_vector}
\alias{check_prob_vector}
\title{Validate probability vector}
\usage{
check_prob_vector(vec, name)
}
\arguments{
\item{vec}{Numeric vector. The probability vector to be checked.}

\item{name}{Character string. The name or label for the vector, used in
error messages.}
}
\value{
Throws an error if the vector is invalid; otherwise, returns nothing.
}
\description{
Checks that values are between 0 and 1 (inclusive), that they sum to 1
(with tolerance of +-0.01).
}
