% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_run_number}
\alias{check_run_number}
\title{Checks if the run number is a non-negative integer.}
\usage{
check_run_number(run_number)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}
}
\value{
Throws an error if the run number is invalid.
}
\description{
Checks if the run number is a non-negative integer.
}
