% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{confidence_interval_method}
\alias{confidence_interval_method}
\title{Use the confidence interval method to select the number of replications.}
\usage{
confidence_interval_method(
  replications,
  desired_precision,
  metric,
  verbose = TRUE
)
}
\arguments{
\item{replications}{Number of times to run the model.}

\item{desired_precision}{Desired mean deviation from confidence interval.}

\item{metric}{Name of performance metric to assess.}

\item{verbose}{Boolean, whether to print messages about parameters.}
}
\value{
Dataframe with results from each replication.
}
\description{
This could be altered to use WelfordStats and ReplicationTabuliser if
desired, but currently is independent.
}
