% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_asu_arrivals}
\alias{create_asu_arrivals}
\title{Acute Stroke Unit (ASU) arrival intervals (days).}
\usage{
create_asu_arrivals(stroke = 1.2, tia = 9.3, neuro = 3.6, other = 3.2)
}
\arguments{
\item{stroke}{Numeric. Mean days between stroke patient arrivals.}

\item{tia}{Numeric. Mean days between transient ischaemic attack (TIA)
patient arrivals.}

\item{neuro}{Numeric. Mean days between complex neurological patient
arrivals.}

\item{other}{Numeric. Mean days between other patient arrivals.}
}
\value{
A named list of arrival intervals for ASU.
}
\description{
For example, a value of 1.2 means a new admission every 1.2 days.
}
