% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_asu_los}
\alias{create_asu_los}
\title{Acute Stroke Unit (ASU) length of stay (LOS) distributions (days).}
\usage{
create_asu_los(
  stroke_noesd_mean = 7.4,
  stroke_noesd_sd = 8.61,
  stroke_esd_mean = 4.6,
  stroke_esd_sd = 4.8,
  stroke_mortality_mean = 7,
  stroke_mortality_sd = 8.7,
  tia_mean = 1.8,
  tia_sd = 2.3,
  neuro_mean = 4,
  neuro_sd = 5,
  other_mean = 3.8,
  other_sd = 5.2
)
}
\arguments{
\item{stroke_noesd_mean}{Numeric. Mean LOS for stroke patients without early
supported discharged (ESD).}

\item{stroke_noesd_sd}{Numeric. SD LOS for stroke patients without ESD.}

\item{stroke_esd_mean}{Numeric. Mean LOS for stroke patients with ESD.}

\item{stroke_esd_sd}{Numeric. SD LOS for stroke patients with ESD.}

\item{stroke_mortality_mean}{Numeric. Mean LOS for stroke patients who pass
away.}

\item{stroke_mortality_sd}{Numeric. SD LOS for stroke patients who pass
away.}

\item{tia_mean}{Numeric. Mean LOS for transient ischemic attack (TIA)
patients.}

\item{tia_sd}{Numeric. SD LOS for TIA patients.}

\item{neuro_mean}{Numeric. Mean LOS for complex neurological patients}

\item{neuro_sd}{Numeric. SD LOS for complex neurological patients}

\item{other_mean}{Numeric. Mean LOS for other patients.}

\item{other_sd}{Numeric. SD LOS for other patients.}
}
\value{
A named list of LOS distributions for ASU.
}
\description{
Mean and standard deviation (SD) of LOS in days in the ASU.
}
