% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_asu_routing}
\alias{create_asu_routing}
\title{ASU routing probabilities.}
\usage{
create_asu_routing(
  stroke_rehab = 0.24,
  stroke_esd = 0.13,
  stroke_other = 0.63,
  tia_rehab = 0.01,
  tia_esd = 0.01,
  tia_other = 0.98,
  neuro_rehab = 0.11,
  neuro_esd = 0.05,
  neuro_other = 0.84,
  other_rehab = 0.05,
  other_esd = 0.1,
  other_other = 0.85
)
}
\arguments{
\item{stroke_rehab}{Numeric. Probability stroke patient to rehab.}

\item{stroke_esd}{Numeric. Probability stroke patient to early supported
discharge (ESD) services.}

\item{stroke_other}{Numeric. Probability stroke patient to other
destinations (e.g., own home, care home, mortality).}

\item{tia_rehab}{Numeric. Probability transient ischemic attack (TIA)
patient to rehab.}

\item{tia_esd}{Numeric. Probability TIA patient to ESD.}

\item{tia_other}{Numeric. Probability TIA patient to other.}

\item{neuro_rehab}{Numeric. Probability complex neurological patient to
rehab.}

\item{neuro_esd}{Numeric. Probability complex neurological patient to ESD.}

\item{neuro_other}{Numeric. Probability complex neurological patient to
other.}

\item{other_rehab}{Numeric. Probability other patient to rehab.}

\item{other_esd}{Numeric. Probability other patient to ESD.}

\item{other_other}{Numeric. Probability other patient to other.}
}
\value{
A named list of routing probabilities for ASU.
}
\description{
Probabilities of each patient type being transferred from the acute
stroke unit (ASU) to other destinations.
}
