% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_asu_trajectory.R
\name{create_asu_trajectory}
\alias{create_asu_trajectory}
\title{Create ASU patient trajectory}
\usage{
create_asu_trajectory(patient_type, param)
}
\arguments{
\item{patient_type}{Character string specifying patient category. Must be
one of: "stroke", "tia", "neuro", or "other". Determines which arrival rate
parameter is used.}

\item{param}{Nested list containing simulation parameters. Must have
structure `param$asu_routing$<patient_type>` containing the probability of
routing to each destination (e.g.`param$asu_routing_stroke$rehab = 0.24`).}
}
\value{
Simmer trajectory object. Defines patient journey logic through the
healthcare system.
}
\description{
Create ASU patient trajectory
}
