% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution_registry.R
\name{create_distribution_registry}
\alias{create_distribution_registry}
\title{Create a distribution registry}
\usage{
create_distribution_registry()
}
\value{
A list containing registry functions and data.
}
\description{
Creates a distribution registry that manages and generates
parameterised samplers for a variety of probability
distributions. Common distributions are included by default,
and more can be added.

Once defined, you can create sampler objects for each
distribution - individually (`dist_create`) or in batches
(`dist_create_batch`) - and then easily draw random samples
from these objects.
}
