% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_parameters}
\alias{create_parameters}
\title{Generate complete parameter list for simulation.}
\usage{
create_parameters(
  asu_arrivals = create_asu_arrivals(),
  rehab_arrivals = create_rehab_arrivals(),
  asu_los = create_asu_los(),
  rehab_los = create_rehab_los(),
  asu_routing = create_asu_routing(),
  rehab_routing = create_rehab_routing(),
  warm_up_period = 365 * 3,
  data_collection_period = 365 * 5,
  number_of_runs = 150,
  scenario_name = NULL,
  cores = 1,
  log_to_console = FALSE,
  log_to_file = FALSE,
  file_path = NULL
)
}
\arguments{
\item{asu_arrivals}{List. Acute stroke unit (ASU) arrival intervals.}

\item{rehab_arrivals}{List. Rehabilitation unit arrival intervals.}

\item{asu_los}{List. ASU length of stay (LOS) distributions.}

\item{rehab_los}{List. Rehabilitation unit LOS distributions.}

\item{asu_routing}{List. ASU routing probabilities.}

\item{rehab_routing}{List. Rehabilitation unit routing probabilities.}

\item{warm_up_period}{Integer. Length of warm-up period (days).}

\item{data_collection_period}{Integer. Length of data collection period
(days).}

\item{number_of_runs}{Integer. Number of simulation runs.}

\item{scenario_name}{Label for scenario (int|float|string).}

\item{cores}{Integer. Number of CPU cores to use.}

\item{log_to_console}{Whether to print activity log to console.}

\item{log_to_file}{Whether to save activity log to file.}

\item{file_path}{Path to save log to file.}
}
\value{
A named list of all simulation parameters.
}
\description{
Generate complete parameter list for simulation.
}
