% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_rehab_arrivals}
\alias{create_rehab_arrivals}
\title{Rehabilitation unit arrival intervals (days).}
\usage{
create_rehab_arrivals(stroke = 21.8, neuro = 31.7, other = 28.6)
}
\arguments{
\item{stroke}{Numeric. Mean days between stroke patient arrivals.}

\item{neuro}{Numeric. Mean days between complex neurological patient
arrivals.}

\item{other}{Numeric. Mean days between other patient arrivals.}
}
\value{
A named list of arrival intervals for rehabilitation unit.
}
\description{
For example, a value of 21.8 means a new admission every 21.8 days.
}
