% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_rehab_los}
\alias{create_rehab_los}
\title{Rehabilitation unit length of stay (LOS) distributions (days).}
\usage{
create_rehab_los(
  stroke_noesd_mean = 28.4,
  stroke_noesd_sd = 27.2,
  stroke_esd_mean = 30.3,
  stroke_esd_sd = 23.1,
  tia_mean = 18.7,
  tia_sd = 23.5,
  neuro_mean = 27.6,
  neuro_sd = 28.4,
  other_mean = 16.1,
  other_sd = 14.1
)
}
\arguments{
\item{stroke_noesd_mean}{Numeric. Mean LOS for stroke patients without early
supported discharged (ESD).}

\item{stroke_noesd_sd}{Numeric. SD LOS for stroke patients without ESD.}

\item{stroke_esd_mean}{Numeric. Mean LOS for stroke patients with ESD.}

\item{stroke_esd_sd}{Numeric. SD LOS for stroke patients with ESD.}

\item{tia_mean}{Numeric. Mean LOS for transient ischemic attack (TIA)
patients.}

\item{tia_sd}{Numeric. SD LOS for TIA patients.}

\item{neuro_mean}{Numeric. Mean LOS for complex neurological patients}

\item{neuro_sd}{Numeric. SD LOS for complex neurological patients}

\item{other_mean}{Numeric. Mean LOS for other patients.}

\item{other_sd}{Numeric. SD LOS for other patients.}
}
\value{
A named list of LOS distributions for rehabilitation unit.
}
\description{
Rehabilitation unit length of stay (LOS) distributions (days).
}
