% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_rehab_routing}
\alias{create_rehab_routing}
\title{Rehabilitation unit routing probabilities.}
\usage{
create_rehab_routing(
  stroke_esd = 0.4,
  stroke_other = 0.6,
  tia_esd = 0L,
  tia_other = 1L,
  neuro_esd = 0.09,
  neuro_other = 0.91,
  other_esd = 0.13,
  other_other = 0.88
)
}
\arguments{
\item{stroke_esd}{Numeric. Probability stroke patient to early supported
discharge (ESD) services.}

\item{stroke_other}{Numeric. Probability stroke patient to other
destinations (e.g., own home, care home, mortality).}

\item{tia_esd}{Numeric. Probability transient ischemic attack (TIA) patient
to ESD.}

\item{tia_other}{Numeric. Probability TIA patient to other.}

\item{neuro_esd}{Numeric. Probability complex neurological patient to ESD.}

\item{neuro_other}{Numeric. Probability complex neurological patient to
other.}

\item{other_esd}{Numeric. Probability other patient to ESD.}

\item{other_other}{Numeric. Probability other patient to other.}
}
\value{
A named list of routing probabilities for rehabilitation unit.
}
\description{
Probabilities of each patient type being transferred from the rehabilitation
unit to other destinations.
}
