% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rehab_trajectory.R
\name{create_rehab_trajectory}
\alias{create_rehab_trajectory}
\title{Create rehab patient trajectory}
\usage{
create_rehab_trajectory(patient_type, param)
}
\arguments{
\item{patient_type}{Character string specifying patient category. Must be
one of: "stroke", "tia", "neuro", or "other". Determines which arrival rate
parameter is used.}

\item{param}{Nested list containing simulation parameters. Must have
structure `param$rehab_routing$<patient_type>` containing the probability of
routing to each destination (e.g.`param$rehab_routing$stroke$esd = 0.40`).}
}
\value{
Simmer trajectory object. Defines patient journey logic through the
healthcare system.
}
\description{
Create rehab patient trajectory
}
