% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_registry.R
\docType{class}
\name{DistributionRegistry}
\alias{DistributionRegistry}
\title{Registry for parametrised probability distributions}
\description{
The DistributionRegistry manages and generates parameterized samplers for a
variety of probability distributions. Common distributions are included by
default, and more can be added.

Once defined, you can create sampler objects for each distribution -
individually (`create`) or in batches (`create_batch`) - and then easily
draw random samples from these objects.

To add more built-in distributions, edit `initialize()`. To add custom ones
at any time, use `register()`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{registry}}{Named list of registered distribution generator functions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DistributionRegistry-new}{\code{DistributionRegistry$new()}}
\item \href{#method-DistributionRegistry-register}{\code{DistributionRegistry$register()}}
\item \href{#method-DistributionRegistry-get}{\code{DistributionRegistry$get()}}
\item \href{#method-DistributionRegistry-transform_to_lnorm}{\code{DistributionRegistry$transform_to_lnorm()}}
\item \href{#method-DistributionRegistry-create}{\code{DistributionRegistry$create()}}
\item \href{#method-DistributionRegistry-create_batch}{\code{DistributionRegistry$create_batch()}}
\item \href{#method-DistributionRegistry-clone}{\code{DistributionRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-new}{}}}
\subsection{Method \code{new()}}{
Pre-registers a set of common base R distribution generators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-register"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-register}{}}}
\subsection{Method \code{register()}}{
Register a distribution generator under a name.

Typically, the generator should be a function that:
1. Accepts parameters for a distribution.
2. Returns another function - the *sampler* - which takes a `size`
argument and produces that many random values from the specified
distribution.

By storing generators rather than fixed samplers, you can create as many
different samplers as you want later, each with different parameters,
while reusing the same generator code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$register(name, generator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name (string)}

\item{\code{generator}}{Function to create a sampler given its parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-get}{}}}
\subsection{Method \code{get()}}{
Get a registered distribution generator by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name (string)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Generator function for the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-transform_to_lnorm"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-transform_to_lnorm}{}}}
\subsection{Method \code{transform_to_lnorm()}}{
Convert mean/sd to lognormal parameters, returning the corresponding
\code{meanlog} and \code{sdlog} parameters used by R's \code{rlnorm()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$transform_to_lnorm(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Named list with two elements: mean and sd.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of the same structure, but with elements
\code{meanlog} and \code{sdlog} for each patient type.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-create"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-create}{}}}
\subsection{Method \code{create()}}{
Create a parameterised sampler for a distribution.

The returned function draws random samples of a specified size from
the given distribution with fixed parameters.

For "lognormal", if "meanlog" and "sdlog" are present in the parameters,
they will be used as-is. If not, but "mean" and "sd" are provided, these
will be transformed into "meanlog"/"sdlog" automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$create(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name}

\item{\code{...}}{Parameters for the generator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A function that draws samples when called.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-create_batch"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-create_batch}{}}}
\subsection{Method \code{create_batch()}}{
Batch-create samplers from a configuration list.

The configuration should be a list of lists, each sublist specifying a
`class_name` (distribution) and `params` (parameter list for that
distribution).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$create_batch(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Named or unnamed list. Each entry is a list with
'class_name' and 'params'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of parameterised samplers (named if config is named).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
