% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occupancy_stats.R
\name{get_occupancy_stats}
\alias{get_occupancy_stats}
\title{Calculate occupancy statistics}
\usage{
get_occupancy_stats(occupancy)
}
\arguments{
\item{occupancy}{DataFrame with three columns: \code{resource}, \code{time},
and \code{occupancy}.}
}
\value{
A list of data frames, one per resource, each containing occupancy
statistics.
}
\description{
For each resource in the data, this function computes how often each
occupancy level was observed. It then calculates:
\itemize{
  \item The frequency of each occupancy level
  \item The proportion (percentage) of time at each occupancy
  \item The cumulative proportion (cumulative percentage) up to each
  occupancy
  \item The probability of delay (probability that occupancy is at or above
  a given level)
  \item The "1 in every n" patients delayed (inverse of probability of delay)
}
}
