% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run simulation.}
\usage{
model(run_number, param, set_seed = TRUE)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}

\item{param}{Named list of model parameters.}

\item{set_seed}{Whether to set seed within the model function (which we
may not wish to do if being set elsewhere - such as done in runner()).
Default is TRUE.}
}
\value{
Named list with three tables: monitored arrivals,
monitored resources, and the processed results from the run.
}
\description{
Run simulation.
}
