% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Generate parameter list for simulation.}
\usage{
parameters(
  parameter_file = NULL,
  parameter_config = NULL,
  warm_up_period = 365L * 3L,
  data_collection_period = 365L * 5L,
  number_of_runs = 150L,
  scenario_name = NULL,
  cores = 1L,
  log_to_console = FALSE,
  log_to_file = FALSE,
  file_path = NULL
)
}
\arguments{
\item{parameter_file}{Path to parameters JSON file. Ignored if
parameter_config given.}

\item{parameter_config}{Optional list of parameters (overrides file).}

\item{warm_up_period}{Integer. Length of warm-up period (days).}

\item{data_collection_period}{Integer. Length of data collection period
(days).}

\item{number_of_runs}{Integer. Number of simulation runs.}

\item{scenario_name}{Label for scenario (int|float|string).}

\item{cores}{Integer. Number of CPU cores to use.}

\item{log_to_console}{Whether to print activity log to console.}

\item{log_to_file}{Whether to save activity log to file.}

\item{file_path}{Path to save log to file.}
}
\value{
A named list of all simulation parameters.
}
\description{
Generate parameter list for simulation.
}
