% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\name{plot_replication_ci}
\alias{plot_replication_ci}
\title{Generate a plot of metric and confidence intervals with increasing
simulation replications}
\usage{
plot_replication_ci(conf_ints, yaxis_title, file_path = NULL, min_rep = NULL)
}
\arguments{
\item{conf_ints}{A dataframe containing confidence interval statistics,
including cumulative mean, upper/lower bounds, and deviations. As returned
by ReplicationTabuliser summary_table() method.}

\item{yaxis_title}{Label for y axis.}

\item{file_path}{Path and filename to save the plot to.}

\item{min_rep}{The number of replications required to meet the desired
precision.}
}
\description{
Generate a plot of metric and confidence intervals with increasing
simulation replications
}
