% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_replications.R
\docType{class}
\name{ReplicationsAlgorithm}
\alias{ReplicationsAlgorithm}
\title{Replication algorithm to automatically select number of replications.}
\value{
Object of `ReplicationsAlgorithm` with methods for determining the
appropriate number of replications to use.
}
\description{
Implements an adaptive replication algorithm for selecting the
appropriate number of simulation replications based on statistical
precision.

Uses the "Replications Algorithm" from Hoad, Robinson, & Davies (2010).
Automated selection of the number of replications for a discrete-event
simulation. Journal of the Operational Research Society.
https://www.jstor.org/stable/40926090.

Given a model's performance measure and a user-set confidence interval
half width prevision, automatically select the number of replications.
Combines the "confidence intervals" method with a sequential look-ahead
procedure to determine if a desired precision in the confidence interval
is maintained.

This class is based on the Python class `ReplicationsAlgorithm` from Tom
Monks (2021) sim-tools: fundamental tools to support the simulation process
in python (https://github.com/TomMonks/sim-tools) (MIT Licence).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{param}}{Model parameters (from parameters()).}

\item{\code{metrics}}{List of performance measure to track (should correspond to
column names from the run results dataframe).}

\item{\code{desired_precision}}{The target half width precision for the algorithm
(i.e. percentage deviation of the confidence interval from the mean,
expressed as a proportion, e.g. 0.1 = 10\%). Choice is fairly arbitrary.}

\item{\code{initial_replications}}{Number of initial replications to perform.}

\item{\code{look_ahead}}{Minimum additional replications to look ahead to assess
stability of precision. When the number of replications is <= 100, the
value of look_ahead is used. When they are > 100, then
look_ahead / 100 * max(n, 100) is used.}

\item{\code{replication_budget}}{Maximum allowed replications. Use for larger
models where replication runtime is a constraint.}

\item{\code{reps}}{Number of replications performed.}

\item{\code{nreps}}{The minimum number of replicatons required to achieve
desired precision for each metric.}

\item{\code{summary_table}}{Dataframe containing cumulative statistics for each
replication for each metric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReplicationsAlgorithm-new}{\code{ReplicationsAlgorithm$new()}}
\item \href{#method-ReplicationsAlgorithm-valid_inputs}{\code{ReplicationsAlgorithm$valid_inputs()}}
\item \href{#method-ReplicationsAlgorithm-klimit}{\code{ReplicationsAlgorithm$klimit()}}
\item \href{#method-ReplicationsAlgorithm-find_position}{\code{ReplicationsAlgorithm$find_position()}}
\item \href{#method-ReplicationsAlgorithm-select}{\code{ReplicationsAlgorithm$select()}}
\item \href{#method-ReplicationsAlgorithm-clone}{\code{ReplicationsAlgorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-new}{}}}
\subsection{Method \code{new()}}{
Initialise the ReplicationsAlgorithm object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$new(
  param,
  metrics = c("mean_waiting_time_nurse", "mean_serve_time_nurse", "utilisation_nurse"),
  desired_precision = 0.1,
  initial_replications = 3L,
  look_ahead = 5L,
  replication_budget = 1000L,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Model parameters.}

\item{\code{metrics}}{List of performance measure to track.}

\item{\code{desired_precision}}{Target half width precision for the algorithm.}

\item{\code{initial_replications}}{Number of initial replications to perform.}

\item{\code{look_ahead}}{Minimum additional replications to look ahead.}

\item{\code{replication_budget}}{Maximum allowed replications.}

\item{\code{verbose}}{Boolean, whether to print messages about parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-valid_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-valid_inputs}{}}}
\subsection{Method \code{valid_inputs()}}{
Checks validity of provided parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$valid_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-klimit"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-klimit}{}}}
\subsection{Method \code{klimit()}}{
Calculate the klimit. Determines the number of additional replications to
check after precision is reached, scaling with total replications if they
are greater than 100. Rounded down to nearest integer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$klimit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Number of additional replications to verify stability (integer).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-find_position"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-find_position}{}}}
\subsection{Method \code{find_position()}}{
Find the first position where element is below deviation, and this is
maintained through the lookahead period.
This is used to correct the ReplicationsAlgorithm, which cannot return
a solution below the initial_replications.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$find_position(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{List of numbers to compare against desired deviation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Integer, minimum replications required to meet and maintain
precision.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-select"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-select}{}}}
\subsection{Method \code{select()}}{
Executes the replication algorithm, determining the necessary number
of replications to achieve and maintain the desired precision.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$select()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplicationsAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReplicationsAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplicationsAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
