% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-package.R
\docType{package}
\name{simulation-package}
\alias{simulation}
\alias{simulation-package}
\title{simulation: Simulation}
\description{
Reproducible analytical pipeline (RAP) for R discrete-event simulation (DES) implementing the Stroke Capacity Planning Model from Monks et al. 2016.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pythonhealthdatascience/rdesrap_stroke}
}

}
\author{
\strong{Maintainer}: Amy Heather \email{a.heather2@exeter.ac.uk}

Other contributors:
\itemize{
  \item Tom Monks \email{t.m.w.monks@exeter.ac.uk} [contributor]
}

}
\keyword{internal}
