% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_warmup.R
\name{time_series_inspection}
\alias{time_series_inspection}
\title{Time series inspection method for determining length of warm-up.}
\usage{
time_series_inspection(result, file_path, warm_up = NULL)
}
\arguments{
\item{result}{Named list with `arrivals` containing output from
`get_mon_arrivals()` and `resources` containing output from
`get_mon_resources()` (`per_resource = TRUE` and `ongoing = TRUE`).}

\item{file_path}{Path to save figure to.}

\item{warm_up}{Location on X axis to plot vertical red line indicating the
chosen warm-up period. Defaults to NULL, which will not plot a line.}
}
\description{
Find the cumulative mean results and plot over time (overall and per run).
}
