% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution_registry.R
\name{transform_to_lnorm}
\alias{transform_to_lnorm}
\title{Transform mean/sd on original scale to meanlog/sdlog for
lognormal}
\usage{
transform_to_lnorm(params)
}
\arguments{
\item{params}{A list with elements `mean` and `sd` (on
original scale).}
}
\value{
A list with elements `meanlog` and `sdlog`.
}
\description{
Transform mean/sd on original scale to meanlog/sdlog for
lognormal
}
