% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{valid_inputs}
\alias{valid_inputs}
\title{Validate input parameters for the simulation.}
\usage{
valid_inputs(run_number, param)
}
\arguments{
\item{run_number}{Integer representing index of current simulation run.}

\item{param}{List containing parameters for the simulation.}
}
\value{
Throws an error if any parameter is invalid.
}
\description{
Validate input parameters for the simulation.
}
