% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_param_names}
\alias{check_param_names}
\title{Validate parameter names.}
\usage{
check_param_names(param)
}
\arguments{
\item{param}{List containing parameters for the simulation.}
}
\value{
Throws an error if there are missing or extra parameters.
}
\description{
Ensure that all required parameters are present, and no extra parameters are
provided.
}
