% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_inputs.R
\name{check_param_values}
\alias{check_param_values}
\title{Validate parameter values.}
\usage{
check_param_values(param)
}
\arguments{
\item{param}{List containing parameters for the simulation.}
}
\value{
Throws an error if any specified parameter value is invalid.
}
\description{
Ensure that specific parameters are positive numbers, or non-negative
integers.
}
