% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rehab_trajectory.R
\name{create_rehab_trajectory}
\alias{create_rehab_trajectory}
\title{Create rehab patient trajectory.}
\usage{
create_rehab_trajectory(env, patient_type, param)
}
\arguments{
\item{env}{Simmer environment object. The simulation environment where
generators will be added.}

\item{patient_type}{Character string specifying patient category. Must be
one of: "stroke", "tia", "neuro", or "other". Determines which arrival rate
parameter is used.}

\item{param}{Nested list containing simulation parameters. Must have
structure \code{param$rehab_routing$<patient_type>} containing the
probability of routing to each destination (e.g.
\code{param$rehab_routing$stroke$esd = 0.40}).}
}
\value{
Simmer trajectory object. Defines patient journey logic through the
healthcare system.
}
\description{
Represents patient stay on the rehabilitation unit - samples their (1)
 destination after rehab, and (2) length of stay (LOS) on the unit.
}
