% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_registry.R
\docType{class}
\name{DistributionRegistry}
\alias{DistributionRegistry}
\title{Registry for parametrised probability distributions}
\description{
The DistributionRegistry manages and generates parameterized samplers for a
variety of probability distributions. Common distributions are included by
default, and more can be added.

Once defined, you can create sampler objects for each distribution -
individually (`create`) or in batches (`create_batch`) - and then easily
draw random samples from these objects.

To add more built-in distributions, edit `initialize()`. To add custom ones
at any time, use `register()`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{registry}}{Named list of registered distribution generator functions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DistributionRegistry-new}{\code{DistributionRegistry$new()}}
\item \href{#method-DistributionRegistry-register}{\code{DistributionRegistry$register()}}
\item \href{#method-DistributionRegistry-get}{\code{DistributionRegistry$get()}}
\item \href{#method-DistributionRegistry-create}{\code{DistributionRegistry$create()}}
\item \href{#method-DistributionRegistry-create_batch}{\code{DistributionRegistry$create_batch()}}
\item \href{#method-DistributionRegistry-clone}{\code{DistributionRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-new}{}}}
\subsection{Method \code{new()}}{
Pre-registers a set of common base R distribution generators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-register"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-register}{}}}
\subsection{Method \code{register()}}{
Register a distribution generator under a name.

Typically, the generator should be a function that takes
distribution-specific parameters and returns a function of `size` (the
sample size).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$register(name, generator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name (string)}

\item{\code{generator}}{Function to create a sampler given its parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-get}{}}}
\subsection{Method \code{get()}}{
Get a registered distribution generator by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name (string)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Generator function for the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-create"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-create}{}}}
\subsection{Method \code{create()}}{
Create a parameterised sampler for a distribution.

The returned function draws random samples of a specified size from
the given distribution with fixed parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$create(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Distribution name}

\item{\code{...}}{Parameters for the generator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A function that draws samples when called.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-create_batch"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-create_batch}{}}}
\subsection{Method \code{create_batch()}}{
Batch-create samplers from a configuration list.

The configuration should be a list of lists, each sublist specifying a
`class_name` (distribution) and `params` (parameter list for that
distribution).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$create_batch(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Named or unnamed list. Each entry is a list with
'class_name' and 'params'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of parameterised samplers (named if config is named).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DistributionRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DistributionRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistributionRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
