% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmup.R
\name{filter_warmup}
\alias{filter_warmup}
\title{Filters arrivals and resources to remove warm-up patients.}
\usage{
filter_warmup(result, warm_up_period)
}
\arguments{
\item{result}{Named list with two tables: monitored arrivals & resources.}

\item{warm_up_period}{Length of warm-up period.}
}
\value{
The name list `result`, but with the tables (`arrivals` and
`resources`) filtered to remove warm-up patients.
}
\description{
Filters arrivals and resources to remove warm-up patients.
}
