% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_routing.R
\name{sample_routing}
\alias{sample_routing}
\title{Sample a destination based on probabilities}
\usage{
sample_routing(prob_list)
}
\arguments{
\item{prob_list}{Named list. The names are destination labels (character),
and the values are their corresponding probabilities (numeric, non-negative,
sum to 1).}
}
\value{
A character string. The name of the selected destination.
}
\description{
Randomly selects a destination from a list, where each destination has an
associated probability.
}
